2048
This repository contains a Python implementation of the popular 2048 board game. The game is played on a 4x4 grid, where the goal is to slide numbered tiles around the board to combine them and create a tile with the value of 2048.

The game has been implemented using object-oriented programming principles and comes with a graphical user interface (GUI) built using the Pygame library. The GUI is designed to be intuitive and easy to use, with colorful tiles and animations to enhance the player experience.

The repository also includes unit tests to ensure that the game logic works correctly, as well as documentation to help new contributors get started with the codebase.

To play the game, simply download the repository and run the main.py file using Python 3. We welcome contributions and bug reports from anyone interested in improving the game or adding new features.

Features:

Object-oriented design
Pygame-based graphical user interface
Game logic unit tests
Documentation
Contributing:

Fork the repository
Make your changes
Create a pull request
License: MIT License

We hope you enjoy playing 2048 as much as we enjoyed building it!